Move custom function tests from square-plugin-example.spec.ts to custom-functions.spec.ts
